fix: reduce CUDA OOM in loophole token ablation
- Add torch.cuda.empty_cache() after each sample to avoid fragmentation
- Load model in fp16 (was fp32) to cut memory (~32GB -> ~16GB for Llama-8B)
- Pin to single GPU (CUDA_VISIBLE_DEVICES=0) to avoid device_map fragmentation
- Set PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True for better allocation

+++ src/utils/model.py
@@ -18,7 +18,7 @@ def load_llm_model_and_tokenizer(model_name):
     tokenizer = AutoTokenizer.from_pretrained(model_name)
     model = AutoModelForCausalLM.from_pretrained(
         model_name,
-        torch_dtype=torch.float32,
+        torch_dtype=torch.float16,
         device_map="auto",
     )
     tokenizer.pad_token = tokenizer.eos_token
